Removing unused features.
I am removing the hint transform because it was primarily used by the
Rekall entities stack. Query rewriting is still a feature I want EFILTER
to offer, but it requires a more thoughtful approach. In absence of a
pressing need for this feature (now that entities are on the backburner)
I am removing the code for now, rather than spending time implementing
it over the new AST types.

BUG=
R=joachim.metz@gmail.com

Review URL: https://codereview.appspot.com/277120043.